JPA
User, Board, Comment

+++ src/main/java/com/yhdc/decrypt/service/UserService.java
@@ -14,73 +14,71 @@ import lombok.RequiredArgsConstructor;
 @Service
 public class UserService {
 	private final UserRepository userRepository;
-	
+
 	// LIST
 	@Transactional(readOnly = true)
 	public List<User> getAllUser() {
 		return userRepository.findAll();
 	}
-	
-	
+
 	// GET
 	@Transactional(readOnly = true)
-	public User getUserById(Long id) {				
-		User userEntity = userRepository.getById(id);		
+	public User getUserById(Long id) {
+		User userEntity = userRepository.getById(id);
 		return userEntity;
 	}
-	
+
 	@Transactional(readOnly = true)
-	public User getUserByUsername(String username) {				
+	public User getUserByUsername(String username) {
 		User userEntity = userRepository.findByUsername(username)
 				.orElseThrow(() -> new IllegalArgumentException("Please check the username again."));
-		
+
 		return userEntity;
 	}
-	
-	
+
 	// NEW
 	@Transactional
-	public User postUser(User user) {		
+	public User postUser(User user) {
 		User newUserEntity = userRepository.save(user);
 		return newUserEntity;
 	}
-	
-	
+
 	// UPDATE username, email, role
 	@Transactional
-	public User putUser(String username, String newUsername, String newEmail, String newRole) {
-		User userEntity = userRepository.findByUsername(username)
-				.orElseThrow(() -> new IllegalArgumentException("Please check the username again."));
-		
-		if (newUsername != null || newUsername != "") {
-			userEntity.setUsername(newUsername);			
+	public User putUser(Long id, String username, String email, String role) {
+		User userEntity = userRepository.getById(id);
+
+		if (username != null || username != "") {
+			userEntity.setUsername(username);
 		}
-		if (newEmail != null || newEmail != "") {
-			userEntity.setEmail(newEmail);			
+		if (email != null || email != "") {
+			userEntity.setEmail(email);
 		}
-		if (newRole!= null || newRole != "") {
-			userEntity.setRole(newRole);			
+		if (role != null || role != "") {
+			userEntity.setRole(role);
 		}
 		return userEntity;
 	}
-	
-	
+
 	// UPDATE password
 	@Transactional
-	public User patchUserPassword(String username, String newPassword) {
-		User userEntity = userRepository.findByUsername(username)
-				.orElseThrow(() -> new IllegalArgumentException("Please check the username again."));
-			userEntity.setPassword(newPassword);		
+	public User patchUserPassword(Long id, String newPassword) {
+		User userEntity = userRepository.getById(id);
+		userEntity.setPassword(newPassword);
 		return userEntity;
 	}
-	
-	
+
 	// DELETE
 	@Transactional
-	public void deleteUser(String username) {
-		User delUser = userRepository.findByUsername(username)
-				.orElseThrow(() -> new IllegalArgumentException("Please check the username again."));
+	public int deleteUser(Long id) {
+		User delUser = userRepository.getById(id);
 
-		userRepository.delete(delUser);
+		if (delUser == null) {
+			return 0;
+		} else {
+			userRepository.delete(delUser);
+		}
+
+		return 1;
 	}
 }
